backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.24.0
python-multipart==0.0.6
email-validator==2.0.0
python-dotenv==1.0.0
pymongo==4.3.3
motor==3.1.1
beanie==1.19.0
